# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EcoLibrary API",
        "description": "API de gestion d'une bibliothèque — catégories, livres, disponibilité et état de la collection.",
        "contact": {
            "email": "admin@ecolibrary.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://project.test/api/v1",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
...
        "/api/livres": {
            "get": {
                "tags": [
                    "Livres"
                ],
                "summary": "Lister tous les livres",
                "operationId": "870f42bc4bad65bdcf897f8c5d56a6de",
                "parameters": [
                    {
                        "name": "categorie_id",
                        "in": "query",
                        "description": "Filtrer par catégorie",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "disponible",
                        "in": "query",
                        "description": "Uniquement les disponibles",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des livres"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Livres"
                ],
                "summary": "Ajouter un livre (admin)",
                "operationId": "eae8b1da3d6e78603a9ceb4a1eb1ed27",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Livre"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Livre créé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/livres/{id}": {
            "get": {
                "tags": [
                    "Livres"
                ],
                "summary": "Afficher un livre",
                "operationId": "9b10c690d5c3817c8f0fa3207d247b1d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail du livre"
                    },
                    "404": {
                        "description": "Livre introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Livres"
                ],
                "summary": "Modifier un livre (admin)",
                "operationId": "2a75c778f322b16ea9a9eb2db0d41224",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Livre"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Livre mis à jour"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Livre introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Livres"
                ],
                "summary": "Supprimer un livre (admin)",
                "operationId": "25191814632cc1504f3e3ac94e3bed87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livre supprimé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Livre introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/livres/recherche": {
            "get": {
                "tags": [
                    "Livres"
                ],
                "summary": "Rechercher un livre par titre ou catégorie",
                "operationId": "595023679598e92b42a2c3fd15a79225",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Mot clé",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultats de recherche"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/livres/populaires": {
            "get": {
                "tags": [
                    "Livres"
                ],
                "summary": "Livres les plus consultés",
                "operationId": "54fa4ddc3fd5a2d35c5d0bf8074934cd",
                "parameters": [
                    {
                        "name": "limite",
                        "in": "query",
                        "description": "Nombre de résultats (défaut 10)",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livres populaires"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/livres/recents": {
            "get": {
                "tags": [
                    "Livres"
                ],
                "summary": "Nouveaux arrivages",
                "operationId": "357b5752806bc3f641d02c8bc9260472",
                "parameters": [
                    {
                        "name": "jours",
                        "in": "query",
                        "description": "Nombre de jours (défaut 30)",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livres récents"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Livre": {
                "required": [
                    "categorie_id",
                    "titre",
                    "auteur"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "categorie_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "titre": {
                        "type": "string",
                        "example": "Le Petit Prince"
                    },
                    "auteur": {
                        "type": "string",
                        "example": "Antoine de Saint-Exupéry"
                    },
                    "isbn": {
                        "type": "string",
                        "example": "978-2-07-040850-4"
                    },
                    "description": {
                        "type": "string",
                        "example": "Un conte poétique et philosophique"
                    },
                    "total_exemplaires": {
                        "type": "integer",
                        "example": 5
                    },
                    "exemplaires_dispo": {
                        "type": "integer",
                        "example": 3
                    },
                    "exemplaires_degrades": {
                        "type": "integer",
                        "example": 1
                    },
                    "nb_consultations": {
                        "type": "integer",
                        "example": 42
                    },
                    "date_publication": {
                        "type": "string",
                        "format": "date",
                        "example": "1943-04-06"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token",
                "description": "Entrez votre token Sanctum : Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}